+++ packages/devtools/client/pages/modules/server-tasks.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import Fuse from 'fuse.js'
-import type { ServerTaskInfo } from '../../../src/types/tasks'
+import type { ServerTaskInfo } from '~/../../src/types'
 import ServerTaskListItem from '~/components/ServerTaskListItem.vue'
 
 definePageMeta({
