+++ packages/devtools/src/types/tasks.ts
@@ -2,11 +2,3 @@ export interface CronCollection {
   cron: string
   tasks: string[]
 }
-
-export interface ServerTaskInfo {
-  name: string
-  handler: string
-  description: string
-  type: 'collection' | 'task'
-  tasks?: ServerTaskInfo[]
-}
